Extract to src/components/managerlib/pdfwatcher.ts
