+++ src/components/manager.ts
@@ -14,6 +14,8 @@ import {Suggestion as CmdEntry} from '../providers/completer/command'
 import {Suggestion as EnvEntry} from '../providers/completer/environment'
 import {Suggestion as GlossEntry} from 'src/providers/completer/glossary'
 
+import {PdfWatcher} from './managerlib/pdfwatcher'
+
 /**
  * The content cache for each LaTeX file `filepath`.
  */
@@ -68,16 +70,14 @@ export class Manager {
 
     private readonly extension: Extension
     private fileWatcher?: chokidar.FSWatcher
-    private pdfWatcher?: chokidar.FSWatcher
     private bibWatcher?: chokidar.FSWatcher
+    private readonly pdfWatcher: PdfWatcher
     private filesWatched: string[] = []
-    private pdfsWatched: string[] = []
     private bibsWatched: string[] = []
     private watcherOptions: chokidar.WatchOptions
     private rsweaveExt: string[] = ['.rnw', '.Rnw', '.rtex', '.Rtex', '.snw', '.Snw']
     private jlweaveExt: string[] = ['.jnw', '.jtexw']
     private weaveExt: string[] = []
-    private pdfWatcherOptions: chokidar.WatchOptions
 
     constructor(extension: Extension) {
         this.extension = extension
@@ -85,7 +85,6 @@ export class Manager {
         const usePolling = configuration.get('latex.watch.usePolling') as boolean
         const interval = configuration.get('latex.watch.interval') as number
         const delay = configuration.get('latex.watch.delay') as number
-        const pdfDelay = configuration.get('latex.watch.pdfDelay') as number
         this.weaveExt = this.jlweaveExt.concat(this.rsweaveExt)
         this.watcherOptions = {
             useFsEvents: false,
@@ -94,14 +93,7 @@ export class Manager {
             binaryInterval: Math.max(interval, 1000),
             awaitWriteFinish: {stabilityThreshold: delay}
         }
-        this.pdfWatcherOptions = {
-            useFsEvents: false,
-            usePolling,
-            interval,
-            binaryInterval: Math.max(interval, 1000),
-            awaitWriteFinish: {stabilityThreshold: pdfDelay}
-        }
-        this.initiatePdfWatcher()
+        this.pdfWatcher = new PdfWatcher(extension)
     }
 
     /**
@@ -842,35 +834,8 @@ export class Manager {
         }
     }
 
-    private initiatePdfWatcher() {
-        if (this.pdfWatcher !== undefined) {
-            return
-        }
-        this.extension.logger.addLogMessage('Creating PDF file watcher.')
-        this.pdfWatcher = chokidar.watch([], this.pdfWatcherOptions)
-        this.pdfWatcher.on('change', (file: string) => this.onWatchedPdfChanged(file))
-        this.pdfWatcher.on('unlink', (file: string) => this.onWatchedPdfDeleted(file))
-    }
-
-    private onWatchedPdfChanged(file: string) {
-        this.extension.logger.addLogMessage(`PDF file watcher - file changed: ${file}`)
-        this.extension.viewer.refreshExistingViewer()
-    }
-
-    private onWatchedPdfDeleted(file: string) {
-        this.extension.logger.addLogMessage(`PDF file watcher - file deleted: ${file}`)
-        if (this.pdfWatcher) {
-            this.pdfWatcher.unwatch(file)
-        }
-        this.pdfsWatched.splice(this.pdfsWatched.indexOf(file), 1)
-    }
-
     watchPdfFile(pdfPath: string) {
-        if (this.pdfWatcher && !this.pdfsWatched.includes(pdfPath)) {
-            this.extension.logger.addLogMessage(`Added to PDF file watcher: ${pdfPath}`)
-            this.pdfWatcher.add(pdfPath)
-            this.pdfsWatched.push(pdfPath)
-        }
+        this.pdfWatcher.watchPdfFile(pdfPath)
     }
 
     private buildOnFileChanged(file: string, bibChanged: boolean = false) {
